requirements.txt
# Core application
fastapi
python-dotenv
filelock
PyYAML

# Web server
gunicorn
uvicorn[standard]

# Logic and APIs
python-binance
python-telegram-bot==13.15
pandas
numpy
requests
matplotlib
mplfinance
psutil

# Version constraints (to prevent dependency conflicts
urllib3<2.0

# SSH Tunneling
sshtunnel
paramiko<3.0

 Firebase
firebase-admin
